requirements.txt
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
astroid==3.2.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
boltons==24.0.0
boto3==1.35.47
botocore==1.35.47
Brotli==1.1.0
bs4==0.0.2
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
courlan==1.3.1
cryptography==43.0.1
dataclasses==0.6
dateparser==1.2.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docutils==0.21.2
exceptiongroup==1.2.2
feedgen==1.0.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Compress==1.15
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Script==2.0.6
Flask-Session==0.8.0
google-api-core==2.22.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-texttospeech==2.21.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
gTTS==2.5.3
gunicorn==23.0.0
h11==0.14.0
htmldate==1.9.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jusText==3.0.1
lazy-object-proxy==1.10.0
lxml==5.3.0
lxml_html_clean==0.3.1
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
msgspec==0.18.6
mutagen==1.47.0
mypy-extensions==1.0.0
nltk==3.9.1
oauthlib==3.2.2
obscure==1.0.1
openai==1.52.0
packaging==24.1
pathspec==0.12.1
pip-review==1.3.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pylint==3.2.7
pymage-size==1.7.1
PyMySQL==1.1.1
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
redis==5.0.8
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
text-unidecode==1.3
tld==0.13
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.66.5
trafilatura==1.12.2
typing_extensions==4.12.2
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.2
zstandard==0.23.0
